Add --batch flag when importing gpg secret key
`gpg --import` asks for a password since version 2.1 when importing a
secret key. Unfortunately, this invokes `pinentry`, which is confused by
the lack of a UI and emits an ominous error, which gpg renders verbatim
without mentioning pinentry:

  "gpg: key ...: error sending to agent: Inappropriate ioctl for device"

This commit adds the --batch flag to gpg when importing a secret key to
import the key unaltered, i.e. with the original passphrase, as
suggested here: https://dev.gnupg.org/T2313