Make gpg_key completely optional
In the current codebase GPG_KEY is set to disabled if gpg_key is unset. This
causes a problem because GPG_KEY is completely optional when GPG_KEYS_ENC and
GPG_KEY_SIGN are in use together.

This change will comment out GPG_KEY instead of adding it with disabled= to
allow GPG_KEYS_ENC and GPG_KEY_SIGN to be used.